Refactor: 피드백 받은 내용 수정(파일이동, 정확한 명칭 사용 등)
