+++ src/app/classroom/[lectureId]/page.tsx
@@ -5,11 +5,11 @@ import React, { FC } from "react";
 import LectureHeader from "@/components/lecture/LectureHeader";
 import LectureVideo from "@/components/lecture/LectureVideo";
 import LectureComment from "@/components/lecture/LectureComment";
-import LectureFooter from "@/components/lecture/LectureFooter";
+import LectureFooter from "@/components/lecture/LectureNavigation";
 
 const LectureHome: FC = () => {
   return (
-    <div className="lectuerContainer flex flex-col w-full h-full">
+    <main className="lectuerContainer flex flex-col w-full h-full">
       <LectureHeader />
       <div className="mainContainer flex w-full h-screen">
         <div className="Container flex flex-col w-3/4">
@@ -18,7 +18,7 @@ const LectureHome: FC = () => {
         </div>
         <LectureComment />
       </div>
-    </div>
+    </main>
   );
 };
