+++ src/components/lecture/LectureComment.tsx
@@ -1,6 +1,6 @@
 import React, { FC } from "react";
 
-import CommentsSection from "@/components/modal/comment/CommentsSection";
+import CommentsSection from "@/components/lecture/CommentsSection";
 import ReplySection from "@/components/modal/comment/ReplySection";
 import { useLectureComment } from "@/hooks/useLectureComment";
 
@@ -8,7 +8,7 @@ const LectureComment: FC = () => {
   const {
     comments,
     activeCommentIndex,
-    isWriteModalOpen,
+    isCommentModalOpen,
     handleComment,
     handleReply,
     setActiveCommentIndex,
@@ -18,8 +18,8 @@ const LectureComment: FC = () => {
   } = useLectureComment();
 
   return (
-    <div className="CommunityContainer bg-gray-100 w-1/4 float-right h-full">
-      <div className=" m-3 flex content-center justify-between items-center">
+    <section className="CommunityContainer bg-gray-100 w-1/4 float-right h-full">
+      <header className=" m-3 flex content-center justify-between items-center">
         <h1 className="text-center text-xl font-semibold pl-2">강의 커뮤니티</h1>
         <button
           className=" m-5 h-11 w-28 text-white rounded-md bg-blue-600 hover:bg-white hover:border hover:border-blue-600 hover:text-blue-600"
@@ -27,13 +27,13 @@ const LectureComment: FC = () => {
         >
           작성
         </button>
-      </div>
-      <div>
+      </header>
+      <main>
         <CommentsSection
           role={"수강생"}
           comments={comments}
           setActiveCommentIndex={setActiveCommentIndex}
-          isWriteModalOpen={isWriteModalOpen}
+          isWriteModalOpen={isCommentModalOpen}
           handleComment={handleComment}
           handleBtn={closeCommentModal}
         />
@@ -46,8 +46,8 @@ const LectureComment: FC = () => {
             handleBtn={closeReplySection}
           />
         )}
-      </div>
-    </div>
+      </main>
+    </section>
   );
 };
